OOP Overhaul
I have added an OOP overhaul / prototype into my codebase. Right now, there is no collection of the event metadata anywhere in the program. Essentially meaning I have no times or events to relate to each other.

I have added a framework that allows for basic event objects to have a start time, end time, name, and description. Do we want to collect this data? If so how? All I have created is a framework to store it. We do this, or I will just post incremental reminders every x minutes (make this time adjustable) that has a generic message.